+++ Samples~/Scripts/LayerAttributeExample.cs
@@ -4,7 +4,7 @@ namespace SaintsField.Samples.Scripts
 {
     public class LayerAttributeExample: MonoBehaviour
     {
-        [Layer] public string layerString;
-        [Layer] public int layerInt;
+        [Layer] public string layerStr;
+        [Layer] public int layerBit;
     }
 }